Update Timber.cpp
The game as it is didn't work for me. I had also tried copying and pasting this code.
Thanks to the suggestion of a friend of mine, I tried to experiment with the accetInput variable and found that this specifically did not make the left and right arrow inputs work.
I am referring to the statement
acceptInput = true;
at line 283 in the block
if (Keyboard::isKeyPressed(Keyboard::Return))
I just added a C style comment next to the statement.